+++ esphome/components/rc522_spi/binary_sensor.py
@@ -1,44 +1,9 @@
-import esphome.codegen as cg
-import esphome.config_validation as cv
-from esphome.components import binary_sensor
-from esphome.const import CONF_UID, CONF_ID
-from esphome.core import HexInt
-from . import rc522_spi_ns, RC522
+import esphome.components.rc522.binary_sensor as rc522_binary_sensor
 
-DEPENDENCIES = ['rc522_spi']
+DEPENDENCIES = ['rc522']
 
-CONF_RC522_ID = 'rc522_id'
-
-
-def validate_uid(value):
-    value = cv.string_strict(value)
-    for x in value.split('-'):
-        if len(x) != 2:
-            raise cv.Invalid("Each part (separated by '-') of the UID must be two characters "
-                             "long.")
-        try:
-            x = int(x, 16)
-        except ValueError as err:
-            raise cv.Invalid("Valid characters for parts of a UID are 0123456789ABCDEF.") from err
-        if x < 0 or x > 255:
-            raise cv.Invalid("Valid values for UID parts (separated by '-') are 00 to FF")
-    return value
-
-
-RC522BinarySensor = rc522_spi_ns.class_('RC522BinarySensor', binary_sensor.BinarySensor)
-
-CONFIG_SCHEMA = binary_sensor.BINARY_SENSOR_SCHEMA.extend({
-    cv.GenerateID(): cv.declare_id(RC522BinarySensor),
-    cv.GenerateID(CONF_RC522_ID): cv.use_id(RC522),
-    cv.Required(CONF_UID): validate_uid,
-})
+CONFIG_SCHEMA = rc522_binary_sensor.CONFIG_SCHEMA
 
 
 def to_code(config):
-    var = cg.new_Pvariable(config[CONF_ID])
-    yield binary_sensor.register_binary_sensor(var, config)
-
-    hub = yield cg.get_variable(config[CONF_RC522_ID])
-    cg.add(hub.register_tag(var))
-    addr = [HexInt(int(x, 16)) for x in config[CONF_UID].split('-')]
-    cg.add(var.set_uid(addr))
+    yield rc522_binary_sensor.to_code(config)
